feat: complete Modules
